Splited ScpEvents in multiple files
